requirements.txt
beautifulsoup4==4.4.1
Flask==0.10.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pycrypto==2.6.1
requests==2.7.0
ujson==1.33
Werkzeug==0.11.10
